Trait test_runner::ChainInfo

pub trait ChainInfo: Sized {
    type Block: BlockT;
    type Executor: NativeExecutionDispatch + 'static;
    type Runtime: Config;
    type RuntimeApi: Send + Sync + 'static + ConstructRuntimeApi<Self::Block, TFullClient<Self::Block, Self::RuntimeApi, Self::Executor>>;
    type SelectChain: SelectChain<Self::Block> + 'static;
    type BlockImport: Send + Sync + Clone + BlockImport<Self::Block, Error = Error, Transaction = TransactionFor<TFullClient<Self::Block, Self::RuntimeApi, Self::Executor>, Self::Block>> + 'static;
    type SignedExtras: SignedExtension;
    type InherentDataProviders: InherentDataProvider + 'static;
    fn signed_extras(
        from: <Self::Runtime as Config>::AccountId
    ) -> Self::SignedExtras;

    fn config(task_executor: TaskExecutor) -> Configuration;

    fn create_client_parts(
        config: &Configuration
    ) -> Result<(Arc<TFullClient<Self::Block, Self::RuntimeApi, Self::Executor>>, Arc<TFullBackend<Self::Block>>, SyncCryptoStorePtr, TaskManager, Box<dyn CreateInherentDataProviders<Self::Block, (), InherentDataProviders = Self::InherentDataProviders>>, Option<Box<dyn ConsensusDataProvider<Self::Block, Transaction = TransactionFor<TFullClient<Self::Block, Self::RuntimeApi, Self::Executor>, Self::Block>>>>, Self::SelectChain, Self::BlockImport), Error>;

    fn dispatch_with_root(
        call: <Self::Runtime as Config>::Call, 
        node: &mut Node<Self>
    );
}

Wrapper trait for concrete type required by this testing framework.

Associated Types

type Block: BlockT

Opaque block type

type Executor: NativeExecutionDispatch + 'static

Executor type

type Runtime: Config

Runtime

type RuntimeApi: Send + Sync + 'static + ConstructRuntimeApi<Self::Block, TFullClient<Self::Block, Self::RuntimeApi, Self::Executor>>

RuntimeApi

type SelectChain: SelectChain<Self::Block> + 'static

select chain type.

type BlockImport: Send + Sync + Clone + BlockImport<Self::Block, Error = Error, Transaction = TransactionFor<TFullClient<Self::Block, Self::RuntimeApi, Self::Executor>, Self::Block>> + 'static

Block import type.

type SignedExtras: SignedExtension

The signed extras required by the runtime

type InherentDataProviders: InherentDataProvider + 'static

The inherent data providers.

Loading content...
Required methods

fn signed_extras(
from: <Self::Runtime as Config>::AccountId
) -> Self::SignedExtras

Signed extras, this function is caled in an externalities provided environment.

fn config(task_executor: TaskExecutor) -> Configuration

config factory

fn create_client_parts(
config: &Configuration
) -> Result<(Arc<TFullClient<Self::Block, Self::RuntimeApi, Self::Executor>>, Arc<TFullBackend<Self::Block>>, SyncCryptoStorePtr, TaskManager, Box<dyn CreateInherentDataProviders<Self::Block, (), InherentDataProviders = Self::InherentDataProviders>>, Option<Box<dyn ConsensusDataProvider<Self::Block, Transaction = TransactionFor<TFullClient<Self::Block, Self::RuntimeApi, Self::Executor>, Self::Block>>>>, Self::SelectChain, Self::BlockImport), Error>

Attempt to create client parts, including block import, select chain strategy and consensus data provider.

fn dispatch_with_root(
call: <Self::Runtime as Config>::Call,
node: &mut Node<Self>
)

Given a call and a handle to the node, execute the call with root privileges.

Loading content...