Trait test_runner::ChainInfo

pub trait ChainInfo: Sized {
    type Block: BlockT;
    type ExecutorDispatch: NativeExecutionDispatch + 'static;
    type Runtime: Config;
    type RuntimeApi: Send + Sync + 'static + ConstructRuntimeApi<Self::Block, TFullClient<Self::Block, Self::RuntimeApi, NativeElseWasmExecutor<Self::ExecutorDispatch>>>;
    type SelectChain: SelectChain<Self::Block> + 'static;
    type BlockImport: Send + Sync + Clone + BlockImport<Self::Block, Error = Error, Transaction = TransactionFor<TFullClient<Self::Block, Self::RuntimeApi, NativeElseWasmExecutor<Self::ExecutorDispatch>>, Self::Block>> + 'static;
    type SignedExtras: SignedExtension;
    type InherentDataProviders: InherentDataProvider + 'static;
    fn signed_extras(
        from: <Self::Runtime as Config>::AccountId
    ) -> Self::SignedExtras;
}

Wrapper trait for concrete type required by this testing framework.

Associated Types

type Block: BlockT

Opaque block type

type ExecutorDispatch: NativeExecutionDispatch + 'static

ExecutorDispatch dispatch type

type Runtime: Config

Runtime

type RuntimeApi: Send + Sync + 'static + ConstructRuntimeApi<Self::Block, TFullClient<Self::Block, Self::RuntimeApi, NativeElseWasmExecutor<Self::ExecutorDispatch>>>

RuntimeApi

type SelectChain: SelectChain<Self::Block> + 'static

select chain type.

type BlockImport: Send + Sync + Clone + BlockImport<Self::Block, Error = Error, Transaction = TransactionFor<TFullClient<Self::Block, Self::RuntimeApi, NativeElseWasmExecutor<Self::ExecutorDispatch>>, Self::Block>> + 'static

Block import type.

type SignedExtras: SignedExtension

The signed extras required by the runtime

type InherentDataProviders: InherentDataProvider + 'static

The inherent data providers.

Required methods

fn signed_extras(
from: <Self::Runtime as Config>::AccountId
) -> Self::SignedExtras

Signed extras, this function is caled in an externalities provided environment.